Refactor low memory test into generic fault injection
