+++ .github/workflows/cicd.yml
@@ -351,18 +351,18 @@ jobs:
       gather_dumps: true
       capture_etw: true
 
-  # Run the low memory simulator in GitHub.
-  low_memory:
+  # Run the fault injection simulator in GitHub.
+  fault_injection:
     needs: regular
     uses: ./.github/workflows/reusable-test.yml
     with:
-      name: low_memory
+      name: fault_injection
       test_command: .\unit_tests.exe
       build_artifact: Build-x64
       environment: windows-2022
       code_coverage: true
       gather_dumps: true
-      low_memory: true
+      fault_injection: true
 
   # Additional jobs to run on a schedule only (skip push and pull request).
   # ---------------------------------------------------------------------------
@@ -375,17 +375,17 @@ jobs:
       build_codeql: true
 
 
-  # Run the complete low memory simulator in GitHub.
+  # Run the complete fault injection simulator in GitHub.
   # Runs on a schedule as this takes a long time to run.
-  low_memory_full:
+  fault_injection_full:
     needs: regular
     if: github.event_name == 'schedule'
     uses: ./.github/workflows/reusable-test.yml
     with:
-      name: low_memory_full
+      name: fault_injection_full
       test_command: .\unit_tests.exe
       build_artifact: Build-x64
       environment: windows-2019
       code_coverage: false
       gather_dumps: true
-      low_memory: true
+      fault_injection: true
