+++ libs/platform/user/platform_user.vcxproj
@@ -30,9 +30,9 @@
     <ClCompile Include="..\ebpf_serialize.c" />
     <ClCompile Include="..\ebpf_state.c" />
     <ClCompile Include="..\ebpf_trampoline.c" />
+    <ClCompile Include="ebpf_fault_injection.cpp" />
     <ClCompile Include="ebpf_handle_user.c" />
     <ClCompile Include="ebpf_leak_detector.cpp" />
-    <ClCompile Include="ebpf_low_memory_test.cpp" />
     <ClCompile Include="ebpf_native_user.c" />
     <ClCompile Include="ebpf_platform_user.cpp" />
     <ClCompile Include="kernel_um.cpp" />
@@ -47,8 +47,8 @@
     <ClInclude Include="..\ebpf_platform.h" />
     <ClInclude Include="..\ebpf_ring_buffer.h" />
     <ClInclude Include="..\ebpf_state.h" />
+    <ClInclude Include="ebpf_fault_injection.h" />
     <ClInclude Include="ebpf_leak_detector.h" />
-    <ClInclude Include="ebpf_low_memory_test.h" />
     <ClInclude Include="ebpf_rundown.h" />
     <ClInclude Include="ebpf_symbol_decoder.h" />
     <ClInclude Include="framework.h" />
