+++ libs/platform/user/platform_user.vcxproj.filters
@@ -82,9 +82,6 @@
     <ClCompile Include="kernel_um.cpp">
       <Filter>Source Files</Filter>
     </ClCompile>
-    <ClCompile Include="ebpf_low_memory_test.cpp">
-      <Filter>Source Files</Filter>
-    </ClCompile>
     <ClCompile Include="ebpf_leak_detector.cpp">
       <Filter>Source Files</Filter>
     </ClCompile>
@@ -92,6 +89,9 @@
       <Filter>Source Files</Filter>
     </ClCompile>
     <ClCompile Include="ebpf_leak_detector.cpp">
+      <Filter>Source Files</Filter>
+    </ClCompile>
+    <ClCompile Include="ebpf_fault_injection.cpp">
       <Filter>Source Files</Filter>
     </ClCompile>
   </ItemGroup>
@@ -132,9 +132,6 @@
     <ClInclude Include="kernel_um.h">
       <Filter>Header Files</Filter>
     </ClInclude>
-    <ClInclude Include="ebpf_low_memory_test.h">
-      <Filter>Header Files</Filter>
-    </ClInclude>
     <ClInclude Include="ebpf_leak_detector.h">
       <Filter>Header Files</Filter>
     </ClInclude>
@@ -142,6 +139,9 @@
       <Filter>Header Files</Filter>
     </ClInclude>
     <ClInclude Include="ebpf_rundown.h">
+      <Filter>Header Files</Filter>
+    </ClInclude>
+    <ClInclude Include="ebpf_fault_injection.h">
       <Filter>Header Files</Filter>
     </ClInclude>
   </ItemGroup>
