+++ scripts/setup_build/setup_build.vcxproj.filters
@@ -67,7 +67,7 @@
     <CopyFileToFolders Include="..\test_execution.json">
       <Filter>Source Files</Filter>
     </CopyFileToFolders>
-    <CopyFileToFolders Include="..\Test-LowMemory.ps1" />
+    <CopyFileToFolders Include="..\Test-FaultInjection.ps1" />
   </ItemGroup>
   <ItemGroup>
     <None Include="..\pre-commit">
