+++ tests/unit/wer_report_test_wrapper.cpp
@@ -41,7 +41,7 @@ AddVectoredExceptionHandler_test(_In_ unsigned long first, _In_ PVECTORED_EXCEPT
 
 unsigned long SetThreadStackGuarantee_test_stack_size_in_bytes = 0;
 
-// Use BOOL to pass "SetThreadStackGuarantee" 
+// Use BOOL to pass "SetThreadStackGuarantee"
 // defined in windows "processthreadapi.h" file
 BOOL
 SetThreadStackGuarantee_test(_Inout_ unsigned long* stack_size_in_bytes)
@@ -50,8 +50,7 @@ SetThreadStackGuarantee_test(_Inout_ unsigned long* stack_size_in_bytes)
     return TRUE;
 }
 
-unsigned long
-WINAPI
+unsigned long WINAPI
 RemoveVectoredExceptionHandler_test(_In_ void* handle)
 {
     UNREFERENCED_PARAMETER(handle);
